# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tabme API",
    "version": "1.0.0",
    "description": "The backend API for the Tabme Chrome Extension. It provides services for user authentication, data storage, and synchronization of spaces, folders, bookmarks, and sticky notes across devices."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Space": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID"
          },
          "title": {
            "type": "string",
            "description": "Space title"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering spaces"
          },
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Folder"
            }
          },
          "widgets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Widget"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Folder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Folder ID"
          },
          "spaceId": {
            "type": "string",
            "format": "uuid",
            "description": "Space ID this folder belongs to"
          },
          "title": {
            "type": "string",
            "description": "Folder title"
          },
          "color": {
            "type": "string",
            "description": "Folder color"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering folders"
          },
          "twoColumn": {
            "type": "boolean",
            "description": "Whether folder displays in two columns"
          },
          "collapsed": {
            "type": "boolean",
            "description": "Whether folder is collapsed"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookmarkItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Bookmark item ID"
          },
          "folderId": {
            "type": "string",
            "format": "uuid",
            "description": "Folder ID this item belongs to"
          },
          "title": {
            "type": "string",
            "description": "Bookmark title"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Bookmark URL"
          },
          "favIconUrl": {
            "type": "string",
            "format": "uri",
            "description": "Favicon URL"
          },
          "position": {
            "type": "string",
            "description": "Position for ordering items"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/items/folder/{folderId}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get items in a folder",
        "description": "Retrieve all bookmark items within a specific folder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "folderId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The folder ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of items in the folder",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookmarkItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Folder not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/items/{itemId}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get item by ID",
        "description": "Retrieve a specific bookmark item by its ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Item details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Update a bookmark item",
        "description": "Update an existing bookmark item",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The bookmark title",
                    "example": "Updated Google"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "The bookmark URL",
                    "example": "https://www.google.com"
                  },
                  "position": {
                    "type": "string",
                    "description": "The position for ordering items",
                    "example": "a1"
                  },
                  "favicon": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL to the bookmark's favicon",
                    "example": "https://www.google.com/favicon.ico"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkItem"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Delete a bookmark item",
        "description": "Delete an existing bookmark item",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Item deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/items": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create a new bookmark item",
        "description": "Create a new bookmark item within a folder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "url",
                  "position",
                  "folderId"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The bookmark title",
                    "example": "Google"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "The bookmark URL",
                    "example": "https://www.google.com"
                  },
                  "position": {
                    "type": "string",
                    "description": "The position for ordering items",
                    "example": "a0"
                  },
                  "folderId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "The folder ID where the item will be created"
                  },
                  "favicon": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL to the bookmark's favicon",
                    "example": "https://www.google.com/favicon.ico"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkItem"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/items/{itemId}/move": {
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Move item to another folder",
        "description": "Move a bookmark item to a different folder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetFolderId",
                  "position"
                ],
                "properties": {
                  "targetFolderId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "The target folder ID"
                  },
                  "position": {
                    "type": "string",
                    "description": "The position in the target folder",
                    "example": "a0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item moved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkItem"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item or folder not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/items/bulk-positions": {
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Bulk update item positions",
        "description": "Update positions of multiple bookmark items at once",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "updates"
                ],
                "properties": {
                  "updates": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "id",
                        "position"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "Item ID"
                        },
                        "position": {
                          "type": "string",
                          "description": "New position"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Positions updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Positions updated successfully"
                    },
                    "updated": {
                      "type": "integer",
                      "description": "Number of items updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/items/{itemId}/mark-used": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Mark item as used",
        "description": "Update the last used timestamp for a bookmark item",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Item marked as used successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Item marked as used"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/items/search": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Search bookmark items",
        "description": "Search for bookmark items by title or URL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query",
            "example": "google"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/BookmarkItem"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "folder": {
                            "type": "object",
                            "properties": {
                              "space": {
                                "$ref": "#/components/schemas/Space"
                              }
                            }
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid search query"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/items/recent": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get recently used items",
        "description": "Retrieve recently used bookmark items for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of recently used items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/BookmarkItem"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "folder": {
                            "type": "object",
                            "properties": {
                              "space": {
                                "$ref": "#/components/schemas/Space"
                              }
                            }
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}